nrf_modem_lib: trace_backends: rtt: early alloc buf
Allocate the modem trace buffer for the RTT backend on boot, instead of
when `nrf_modem_lib_init` is called. This allows the user to delay the
call to `nrf_modem_lib_init`, giving them a chance to attach to the RTT
channel before the modem starts.

Without this change the start of trace logs from the RTT backend are
invariably lost, as the user cannot attach until after data generation
has started.

Signed-off-by: Jordan Yates <jordan@embeint.com>